+++ src/main/tools/window_manage/exposers.ts
@@ -7,7 +7,16 @@
 import { ipcMain } from "electron";
 
 import { openSettingsWindow } from './settings_window';
+import { windowInstance } from './instance';
 
 ipcMain.handle('windowmgr:open:settings', function () {
     return openSettingsWindow();
+});
+
+ipcMain.handle('windowmgr:signal:trigger:refresh', function () {
+    windowInstance.sendToAllWindows({
+        includeMain: true,
+        channel: 'windowmgr:signal:refresh',
+        props: [],
+    });
 });