+++ src/renderer/pages/home/page.tsx
@@ -1,7 +1,9 @@
 // Components
 import { useEffect, useState } from 'react';
 import { Container, FlexDiv } from 'renderer/components/container';
+import { Background } from 'renderer/components/general/background';
 import { CFTextIcon } from 'renderer/components/icons/codeforces';
+import { GoogleIcon } from 'renderer/components/icons/gicon';
 
 // Page Parts
 import { CFContestsView } from './contests_view';
@@ -17,58 +19,54 @@ import { openNewWindowConfig } from 'main/tools/window_opener';
 export function HomePage() {
 
     return (
-        <div className={classNames(
-            'flex',
-            'h-screen w-screen',
-            'text-black dark:text-white',
+        <Background className={classNames(
+            'flex-col'
         )}>
+            {/* Nav Bar */}
+            <NavBar></NavBar>
             <FlexDiv
+                expand={true}
                 className={classNames(
-                    'flex-col',
-                    'bg-bgcolor dark:bg-bgcolor-dark'
-                )}
-                expand={true}>
-                {/* Nav Bar */}
-                <Container
-                    id='topNavBar'
-                    rounded={false}
-                    className={
-                        classNames(
-                            'flex-none',
-                        )}>
-                    <FlexDiv className={classNames(
-                        'w-full',
-                        'flex-row justify-between items-center',
-                        'px-5 py-2'
+                    'flex-col justify-start items-center',
+                )}>
+                <CFContestsView></CFContestsView>
+            </FlexDiv>
+        </Background>
+    );
+}
+
+/**
+ * Navbar component for home page
+ */
+export function NavBar() {
+    return (
+        <Container
+            id='topNavBar'
+            rounded={false}
+            className={
+                classNames(
+                    'flex-none',
+                )}>
+            <FlexDiv className={classNames(
+                'w-full',
+                'flex-row justify-between items-center',
+                'px-5 py-2'
+            )}>
+                <div className='text-2xl'>
+                    <CFTextIcon />
+                </div>
+                {/* Settings Button */}
+                <button
+                    onClick={function () {
+                        window.electron.ipcRenderer.invoke('windowmgr:open:settings');
+                    }}>
+                    <div className={classNames(
+                        'flex flex-none justify-center items-center',
                     )}>
-                        <div className='text-2xl'>
-                            <CFTextIcon />
-                        </div>
-                        <button
-                            onClick={function () {
-                                let param: openNewWindowConfig = {
-                                    url: '/contest',
-                                    isFullUrl: false,
-                                };
-                                console.log('Contact ipcMain to open new window');
-                                window.electron.ipcRenderer.invoke('windowopt:open', param);
-                            }}>
-                            Window.open
-                        </button>
-                        <Link to='/contest' target=''>LinkTo</Link>
-                        <Link to='contest' target='_blank'>LinkTo_blank</Link>
-                        <a href='/contest' target=''>aHref</a>
-                        <a href='/contest' target='_blank'>aHref_blank</a>
-                    </FlexDiv>
-                </Container>
-                <FlexDiv
-                    expand={true}
-                    className={classNames(
-                        'flex-col justify-start items-center',
-                    )}>
-                    <CFContestsView></CFContestsView>
-                </FlexDiv>
-            </FlexDiv >
-        </div >
+                        <GoogleIcon>settings</GoogleIcon>
+                    </div>
+                </button>
+            </FlexDiv>
+        </Container>
     );
 }