+++ src/renderer/pages/settings/settings_data.tsx
@@ -1,0 +1,15 @@
+
+export class SettingsCategoryData {
+    constructor(iconName: string, title: string) {
+        this.iconName = iconName;
+        this.title = title;
+    }
+    iconName: string;
+    title: string;
+}
+
+export let settingCategoryList: SettingsCategoryData[] = [
+    new SettingsCategoryData('palette', 'Appearance'),
+    new SettingsCategoryData('account_circle', 'Account'),
+    new SettingsCategoryData('info', 'About'),
+];