+++ src/renderer/tools/darkmode.tsx
@@ -3,6 +3,9 @@
 // Fundamentals
 import { useLayoutEffect } from 'react';
 
+/**
+ * Returns the system default darkmode pref
+ */
 export function getSystemDarkmodePref(): boolean {
     if (window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches) {
         console.log('Returning true');
@@ -11,13 +14,31 @@ export function getSystemDarkmodePref(): boolean {
     return false;
 }
 
+/**
+ * Set the darkmode of current UI
+ * 
+ * Notice:
+ * - If `darkmode` received `undefined`, then setting the darkmode 
+ * based on system pref
+ */
+export function setUIDarkMode(darkmode?: boolean) {
+    if (darkmode === null) {
+        darkmode = getSystemDarkmodePref();
+    }
+    if (darkmode === true) {
+        document.documentElement.classList.add('dark');
+    }
+    else {
+        document.documentElement.classList.remove('dark');
+    }
+}
 
 /**
  * Update darkmode UI based on the info in the storage and the user system pref
  */
 export async function updateDarkModeUIFromStorage() {
     let curDarkmode = await window.electron.ipcRenderer.invoke('storage:eleCfConfig:getInfo', 'theme.darkmode');
-    if (curDarkmode === undefined) {
+    if (curDarkmode === null) {
         curDarkmode = getSystemDarkmodePref();
     }
     if (curDarkmode === true) {