+++ CHANGELOG.md
@@ -1,6 +1,14 @@
+
+- WorkOn: Settings page.
+- WorkOn: Add account store
+
+## 0.1.1
+
+- Update: Add `Background` component
+
 ## 0.1.0
 
-- WorkOn: Update: Add window management module
-- WorkOn: Update: Add settings page
+- Update: Add window management module
+- Update: Add settings page
 - Update: Add Readme for `EleCFStorage`
 - Update: Change config structure of window management module